refactor system notification to dedicated file
